feat: update field check regex
