+++ packages/console/package.json
@@ -20,6 +20,7 @@
     "@fontsource/roboto-mono": "^4.5.7",
     "@logto/phrases": "^0.1.0",
     "@logto/react": "^0.1.5",
+    "@logto/shared": "^0.1.0",
     "@logto/schemas": "^0.1.0",
     "@mdx-js/react": "^1.6.22",
     "@parcel/core": "^2.5.0",
