--- packages/console/src/utilities/regex.ts
@@ -1,3 +1,0 @@
-// TODO - LOG-1876: Share Regex Between Logto Core and Front-End
-export const emailRegEx = /^\S+@\S+\.\S+$/;
-export const phoneRegEx = /^\d+$/;
