+++ packages/core/package.json
@@ -31,6 +31,7 @@
     "@logto/connector-wechat": "^0.1.0",
     "@logto/connector-wechat-native": "^0.1.0",
     "@logto/phrases": "^0.1.0",
+    "@logto/shared": "^0.1.0",
     "@logto/schemas": "^0.1.0",
     "@silverhand/essentials": "^1.1.6",
     "argon2": "^0.28.5",
