+++ packages/shared/src/index.ts
@@ -1,9 +1,2 @@
-import { existsSync, readFileSync } from 'fs';
-
-export const getFileContents = (filePath: string, fallbackContent: string): string => {
-  if (existsSync(filePath)) {
-    return readFileSync(filePath, 'utf8');
-  }
-
-  return fallbackContent;
-};
+export * from './file-utils';
+export * from './regex';
