+++ packages/shared/src/regex.ts
@@ -1,7 +1,7 @@
 export const emailRegEx = /^\S+@\S+\.\S+$/;
 export const phoneRegEx = /^\d+$/;
-export const usernameRegEx = /^.{3,}$/;
-export const nameRegEx = /^.{3,}$/;
+export const usernameRegEx = /^[A-Z_a-z-][\w-]*$/;
+export const nameRegEx = /^.+$/;
 export const passwordRegEx = /^.{6,}$/;
 export const redirectUriRegEx = /^https?:\/\//;
-export { hexColorRegEx } from '@logto/schemas';
+export const hexColorRegEx = /^#[\da-f]{3}([\da-f]{3})?$/i;
